pyproject.toml
[tool.poetry]
name = "mini-app-demo"
version = "0.1.0"
description = ""
authors = ["Mikhail <mi.sidoren@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytelegrambotapi = "^4.23.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.0"
requests = "^2.32.3"
openai = "^1.51.2"
uvicorn = "^0.31.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
